Refatora disposicao de alguns components
